requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
flake8>=3.6.0,<3.7.0
djongo>=1.2.31,<1.3.0
django-multiselectfield>=0.1.8,<0.2.0
